Split classes into their own respective java files
