+++ src/main/java/Deadline.java
@@ -1,0 +1,13 @@
+public class Deadline extends Task {
+    protected String by;
+
+    public Deadline(String name, boolean isMarked, String by) {
+        super(name, isMarked);
+        this.by = by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by + ")";
+    }
+}
