+++ src/main/java/ToDo.java
@@ -1,0 +1,11 @@
+public class ToDo extends Task {
+
+    public ToDo(String name, boolean isMarked) {
+        super(name, isMarked);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
